+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+V1.0.2 - 1/14/2026
+- Fixed UI for some PowerMob values not updating max range when doing single values.
+- Fixed teleport ability to limit the Y range to be within the defined max distance.
+
 V1.0.1 - 1/11/2026
 - UI item selection is now sorted by id
 - Added more default items.